requirements.txt
Flask==2.3.3
SQLAlchemy==1.4.52
APScheduler==3.11.0
requests==2.31.0
beautifulsoup4==4.12.2
pydantic==1.10.12
python-dotenv==1.0.0
flask-restx==1.3.0
flask-cors==4.0.0
